backend/web-back/requirements.txt
asgiref==3.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
python-dotenv==0.13.0
pytz==2019.3
sqlparse==0.3.1
Django==3.2.5
djangorestframework==3.12.0
djoser==0.4.1
django-cors-headers==3.7.0
mysqlclient==2.0.3
Pillow==8.2.0
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0
django-cleanup==5.2.0
